# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API IBGE BCode",
        "version": "0.1"
    },
    "paths": {
        "/api/ibge": {},
        "/api/ibge/municipios": {
            "get": {
                "operationId": "65e6d3e2ddc9fc70db691cbd02e804d8",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "412 e 500": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ibge/municipios/{ibge}": {
            "get": {
                "operationId": "d47f297917b5e9cd3b2f114fc12ed264",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "412 e 500": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ibge/municipios/uf/{uf}": {
            "get": {
                "operationId": "f439681df6e01f2ba5a69b815d58cd7c",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "412 e 500": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ibge/uf": {
            "get": {
                "operationId": "0a1d37921f3b5fedd6919f1120d2247a",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "412 e 500": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ibge/uf/{code}": {
            "get": {
                "operationId": "85ceeee0b4c7052b956fb417f0a66e2d",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "412 e 500": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "json",
                                        "value": {
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}